requirements.txt
Flask==3.0.2
flask-cors==4.0.0
google-genai==0.3.0
gunicorn==21.2.0